+++ CookieShop.App/DTOs/Cart/CartResponse.cs
@@ -3,5 +3,5 @@ namespace CookieShop.App.DTOs.Cart;
 public record CartResponse(
     Guid Id,
     IReadOnlyCollection<CartItemResponse> Items,
-    decimal TotalAmount
+    long TotalAmount
 );
