+++ CookieShop.App/DTOs/Order/OrderItemResponse.cs
@@ -3,7 +3,7 @@ namespace CookieShop.App.DTOs.Order;
 public record OrderItemResponse(
     Guid ProductId,
     string ProductName,
-    decimal UnitPrice,
+    long UnitPrice,
     int Quantity,
-    decimal LineTotal
+    long LineTotal
 );
