+++ CookieShop.App/DTOs/Order/OrderResponse.cs
@@ -4,7 +4,7 @@ public record OrderResponse(
     Guid Id,
     Guid UserId,
     string Status,
-    decimal TotalAmount,
+    long TotalAmount,
     string? Note,
     DateTimeOffset CreatedAt,
     DateTimeOffset UpdatedAt
